requirements.txt
PyQt5>=5.15
PyMuPDF>=1.22.0
Pillow>=9.0.0
pytest>=7.0
mypy>=0.991

